refactor: consolidate price utilities into utils.ts
- Move all price-related functions from price-utils.ts to utils.ts
- Update imports in TronContractDeployer and register-facets-to-diamond
- Delete the now-redundant price-utils.ts file
- Simplify module structure by reducing file count
